refactor: extract delegates from JointedModelColladaImporter (688→252 lines)
Extract three package-private delegates from JointedModelColladaImporter:
- ColladaImportMeshBuilder: mesh/geometry/weight conversion (196 lines)
- ColladaImportMaterialLoader: texture/material loading (166 lines)
- ColladaImportDebugPrinter: debug print utilities (72 lines)

Public API (constructor, loadSkeletonVisual, loadAliceModel) unchanged.
Characterization tests verify API contracts and matrix conversion behavior.

Closes #617
