+++ core/model-loading/src/main/java/org/lgna/story/resourceutilities/JointedModelColladaImporter.java
@@ -1,27 +1,9 @@
 package org.lgna.story.resourceutilities;
 
 import com.dddviewr.collada.Collada;
-import com.dddviewr.collada.controller.Controller;
-import com.dddviewr.collada.controller.Skin;
-import com.dddviewr.collada.controller.VertexWeights;
-import com.dddviewr.collada.effects.Effect;
-import com.dddviewr.collada.effects.NewParam;
-import com.dddviewr.collada.geometry.Geometry;
-import com.dddviewr.collada.geometry.Primitives;
-import com.dddviewr.collada.geometry.Triangles;
-import com.dddviewr.collada.images.Image;
-import com.dddviewr.collada.materials.InstanceEffect;
-import com.dddviewr.collada.materials.LibraryMaterials;
-import com.dddviewr.collada.materials.Material;
 import com.dddviewr.collada.nodes.Node;
 import com.dddviewr.collada.visualscene.*;
-import com.jogamp.common.nio.Buffers;
-import edu.cmu.cs.dennisc.image.ImageUtilities;
-import edu.cmu.cs.dennisc.print.PrintUtilities;
 import edu.cmu.cs.dennisc.scenegraph.*;
-import edu.cmu.cs.dennisc.scenegraph.Component;
-import edu.cmu.cs.dennisc.texture.BufferedImageTexture;
-import edu.cmu.cs.dennisc.texture.Texture;
 import org.alice.math.immutable.*;
 import org.alice.math.immutable.AngleInDegrees;
 import org.lgna.story.implementation.JointedModelImp.VisualData;
@@ -29,13 +11,10 @@ import org.lgna.story.resources.ImplementationAndVisualType;
 import org.lgna.story.resources.JointedModelResource;
 import org.xml.sax.SAXException;
 
-import java.awt.*;
-import java.awt.image.BufferedImage;
 import java.io.File;
 import java.io.IOException;
-import java.util.*;
+import java.util.ArrayList;
 import java.util.List;
-import java.util.Map.Entry;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
@@ -108,26 +87,11 @@ public class JointedModelColladaImporter {
     if (rootNode != null) {
       return rootNode;
     }
-    //No joints are found, so return null
     return null;
   }
 
-  private AffineMatrix4x4 floatArrayToAliceMatrix(float[] floatData) throws ModelLoadingException {
-    double[] doubleData = new double[floatData.length];
-    for (int i = 0; i < floatData.length; i++) {
-      doubleData[i] = floatData[i];
-    }
-    AffineMatrix4x4 srcMatrix;
-    if (doubleData.length == 12 || doubleData.length == 16) {
-      srcMatrix = AffineMatrix4x4.createFromRowMajorArray(doubleData);
-    } else {
-      throw new ModelLoadingException("Error converting collada matrix to Alice matrix. Expected array of size 12 or 16, instead got " + floatData.length);
-    }
-    return orientation.orientMatrixToAlice(srcMatrix);
-  }
-
   private AffineMatrix4x4 colladaMatrixToAliceMatrix(Matrix m) throws ModelLoadingException {
-    return floatArrayToAliceMatrix(m.getData());
+    return ColladaImportMeshBuilder.floatArrayToAliceMatrix(m.getData(), orientation);
   }
 
   private Joint createAliceSkeletonFromNode(Node node) throws ModelLoadingException {
@@ -184,356 +148,10 @@ public class JointedModelColladaImporter {
     return aliceMatrix;
   }
 
-  private static int getMaterialIndex(String materialId, Collada colladaModel) {
-    if (materialId == null) {
-      return -1;
-    }
-    int index = 0;
-    for (Material material : colladaModel.getLibraryMaterials().getMaterials()) {
-      if (materialId.equals(material.getId())) {
-        return index;
-      }
-      index++;
-    }
-    return -1;
-  }
-
-  private static Controller getControllerForGeometry(Geometry geometry, Collada colladaModel) {
-    if (colladaModel.getLibraryControllers() != null && colladaModel.getLibraryControllers().getControllers() != null) {
-      for (Controller controller : colladaModel.getLibraryControllers().getControllers()) {
-        Geometry foundGeometry = colladaModel.findGeometry(controller.getSkin().getSource());
-        if (foundGeometry == geometry) {
-          return controller;
-        }
-      }
-    }
-    return null;
-  }
-
-  private WeightInfo createWeightInfoForController(Controller meshController) throws ModelLoadingException {
-    /*
-     *  Here is an example of a more complete <vertex_weights> element. Note that the <vcount> element
-      says that the first vertex has 3 bones, the second has 2, etc. Also, the <v> element says that the first
-      vertex is weighted with weights[0] towards the bind shape, weights[1] towards bone 0, and
-      weights[2] towards bone 1:
-     *  From the Collada 1.4.1 spec: (Alice, MAYA, and Blender use the 1.4.1 schema for export)
-      <controller id="skin">
-        <skin source="#base_mesh">
-          <source id="Joints">
-            <Name_array count="4"> Root Spine1 Spine2 Head </Name_array>
-            ...
-          </source>
-          <source id="Weights">
-            <float_array count="4"> 0.0 0.33 0.66 1.0 </float_array>
-            ...
-          </source>
-          <source id="Inv_bind_mats">
-            <float_array count="64"> ... </float_array>
-            ...
-          </source>
-          <joints>
-            <input semantic="JOINT" source="#Joints"/>
-            <input semantic="INV_BIND_MATRIX" source="#Inv_bind_mats"/>
-          </joints>
-          <vertex_weights count="4">
-            <input semantic="JOINT" source="#Joints"/>
-            <input semantic="WEIGHT" source="#Weights"/>
-            <vcount>3 2 2 3</vcount>
-            <v>
-            -1 0 0 1 1 2
-            -1 3 1 4
-            -1 3 2 4
-            -1 0 3 1 2 2
-            </v>
-          </vertex_weights>
-        </skin>
-      </controller>
-     */
-    Skin skin = meshController.getSkin();
-    Map<Integer, float[]> jointWeightMap = getJointWeightMap(skin);
-    //Take the weight data and convert it to WeightInfo
-    String[] jointData = skin.getJointData();
-    if (jointData == null) {
-      throw new ModelLoadingException("Error converting mesh " + meshController.getName() + ", no joint data found on mesh skin.");
-    }
-
-    //Array of inverse bind matrices for all the referenced joints. Is indexed into by the joint index.
-    float[] inverseBindMatrixData = skin.getInvBindMatrixData();
-    if (inverseBindMatrixData == null) {
-      throw new ModelLoadingException("Error converting mesh " + meshController.getName() + ", no inverse bind matrix data found on mesh skin.");
-    }
-    WeightInfo weightInfo = new WeightInfo();
-    for (Entry<Integer, float[]> jointAndWeights : jointWeightMap.entrySet()) {
-      int jointIndex = jointAndWeights.getKey();
-      String jointId = jointData[jointIndex];
-      // The Inverse Bind Matrix for jointIndex i. IBMi in the Collada spec.
-      float[] inverseBindMatrix = Arrays.copyOfRange(inverseBindMatrixData, 16 * jointIndex, 16 * jointIndex + 16);
-      AffineMatrix4x4 aliceInverseBindMatrix = floatArrayToAliceMatrix(inverseBindMatrix);
-      InverseAbsoluteTransformationWeightsPair iawp = InverseAbsoluteTransformationWeightsPair.createInverseAbsoluteTransformationWeightsPair(jointAndWeights.getValue(), aliceInverseBindMatrix);
-      if (iawp != null) {
-        weightInfo.addReference(jointId, iawp);
-      }
-    }
-    return weightInfo;
-  }
-
-  private static Map<Integer, float[]> getJointWeightMap(Skin skin) {
-    VertexWeights vertexWeights = skin.getVertexWeights();
-    float[] weightData = skin.getWeightData();
-    int[] vertexWeightData = vertexWeights.getData();
-    Map<Integer, float[]> jointWeightMap = new HashMap<>();
-    int entryCount = 0;
-    //Loop through the vertex weights and build a map of joints (stored as their indices to the joint data) to
-    // arrays of weight info.
-    //The weight info is stored as arrays of weights where the index of the entry is the index of the joint and
-    // the weight values are the weighting of that vertex to the joint
-    for (int vertex = 0; vertex < vertexWeights.getCount(); vertex++) {
-      int boneCount = vertexWeights.getVcount().getData()[vertex];
-      for (int bone = 0; bone < boneCount; bone++) {
-        int jointIndex = vertexWeightData[entryCount * 2];
-        int weightIndex = vertexWeightData[entryCount * 2 + 1];
-
-        float[] weightArray;
-        if (jointWeightMap.containsKey(jointIndex)) {
-          weightArray = jointWeightMap.get(jointIndex);
-        } else {
-          weightArray = new float[vertexWeights.getCount()];
-          jointWeightMap.put(jointIndex, weightArray);
-        }
-        weightArray[vertex] = weightData[weightIndex];
-        entryCount++;
-      }
-    }
-    return jointWeightMap;
-  }
-
-  private Mesh createAliceSGMeshFromGeometry(Geometry geometry, Collada colladaModel) throws ModelLoadingException {
-
-    Controller meshController = getControllerForGeometry(geometry, colladaModel);
-    Mesh sgMesh;
-    if (meshController != null) {
-      sgMesh = new WeightedMesh();
-    } else {
-      sgMesh = new Mesh();
-    }
-    sgMesh.setName(geometry.getName());
-    final float[] normals = geometry.getMesh().getNormalData();
-    if (normals == null) {
-      throw new ModelLoadingException("No normal data found in model.");
-    }
-    orientation.orientNormals(normals, sgMesh.normalBuffer);
-
-    float[] colladaVertices = geometry.getMesh().getPositionData();
-    double[] doubleVertexData = orientation.orientVertices(colladaVertices, sgMesh.vertexBuffer);
-    final float[] coordData = geometry.getMesh().getTexCoordData();
-    if (coordData == null) {
-      throw new ModelLoadingException("No texture coordinate data found in model.");
-    }
-    sgMesh.textCoordBuffer.setValue(Buffers.newDirectFloatBuffer(coordData));
-
-    //Find the triangle data and use it to set the index data
-    Triangles tris = null;
-    for (Primitives p : geometry.getMesh().getPrimitives()) {
-      if (p instanceof Triangles triangles) {
-        if (tris == null) {
-          tris = triangles;
-        } else {
-          modelLoadingLogger.log(Level.WARNING, "Converting mesh '" + geometry.getName() + "': Unsupported primitive count: Found extra triangle primitives, only processing the first.");
-        }
-      } else {
-        modelLoadingLogger.log(Level.WARNING, "Converting mesh '" + geometry.getName() + "': Unsupported primitive type " + p.getClass() + ". Skipping this geometry.");
-      }
-    }
-    if (tris == null) {
-      modelLoadingLogger.log(Level.WARNING, "Error converting mesh " + geometry.getName() + ". No triangle primitive data found. Skipping entire mesh.");
-      return null;
-    }
-    sgMesh.indexBuffer.setValue(Buffers.newDirectIntBuffer(tris.getData()));
-    // TODO Remove this early binding. The material on a piece of geometry references an instance_material by symbol.
-    // The instance_material, which is specific to an instance_controller on a node, uses target to reference the
-    // material in library_materials by id. The following line bypasses the instance_material, requiring it to have
-    // identical symbol and target and forcing all nodes to use the same material.
-    // TODO Stop using the index as the ID.
-    sgMesh.textureId.setValue(getMaterialIndex(tris.getMaterial(), colladaModel));
-
-    if (sgMesh instanceof WeightedMesh mesh) {
-      recordWeights(mesh, meshController, doubleVertexData);
-    }
-    return sgMesh;
-  }
-
-  private void recordWeights(WeightedMesh sgMesh, Controller meshController, double vertices[]) throws ModelLoadingException {
-    //Since this is a weighted mesh, we need to transform the mesh data into the bind space
-    float[] bindMatrixData = meshController.getSkin().getBindShapeMatrix();
-    if (bindMatrixData != null) {
-      AffineMatrix4x4 bindMatrix = floatArrayToAliceMatrix(bindMatrixData);
-      double[] bindSpaceVertices = new double[vertices.length];
-      for (int i = 0; i < vertices.length; i += 3) {
-        bindMatrix.transformPoint3(bindSpaceVertices, i, vertices, i);
-      }
-      sgMesh.vertexBuffer.setValue(Buffers.newDirectDoubleBuffer(bindSpaceVertices));
-    }
-    sgMesh.weightInfo.setValue(createWeightInfoForController(meshController));
-  }
-
-  private List<Mesh> createAliceMeshesFromCollada(Collada colladaModel) throws ModelLoadingException {
-    List<Geometry> geometries = colladaModel.getLibraryGeometries().getGeometries();
-    List<Mesh> meshes = new ArrayList<Mesh>();
-    for (Geometry geometry : geometries) {
-      Mesh mesh = createAliceSGMeshFromGeometry(geometry, colladaModel);
-      if (mesh != null) {
-        meshes.add(mesh);
-      }
-    }
-    return meshes;
-  }
-
   public static SkeletonVisual loadAliceModel(JointedModelResource resource) {
     VisualData<JointedModelResource> v = ImplementationAndVisualType.ALICE.getFactory(resource).createVisualData();
     SkeletonVisual sv = (SkeletonVisual) v.getSgVisuals()[0];
     return sv;
-  }
-
-  private static Texture getAliceTexture(BufferedImage image) {
-    boolean flipImage = true;
-    BufferedImageTexture aliceTexture = new BufferedImageTexture();
-    BufferedImage tex;
-    if (flipImage) {
-      try {
-        int type;
-        if (image.getTransparency() == BufferedImage.OPAQUE) {
-          type = BufferedImage.TYPE_3BYTE_BGR;
-        } else {
-          type = BufferedImage.TYPE_4BYTE_ABGR;
-        }
-        tex = new BufferedImage(image.getWidth(), image.getHeight(), type);
-      } catch (IllegalArgumentException e) {
-        e.printStackTrace();
-        return null;
-      }
-
-      int imageWidth = image.getWidth();
-      int[] tmpData = new int[imageWidth];
-      int row = 0;
-      for (int y = image.getHeight() - 1; y >= 0; y--) {
-        image.getRGB(0, row++, imageWidth, 1, tmpData, 0, imageWidth);
-        tex.setRGB(0, y, imageWidth, 1, tmpData, 0, imageWidth);
-      }
-
-    } else {
-      tex = image;
-    }
-    aliceTexture.setBufferedImage(tex);
-    return aliceTexture;
-  }
-
-  private List<TexturedAppearance> createAliceMaterialsFromCollada(Collada colladaModel, File rootPath, List<Mesh> aliceMeshes) throws ModelLoadingException {
-    List<TexturedAppearance> textureAppearances = new LinkedList<TexturedAppearance>();
-    final LibraryMaterials libraryMaterials = colladaModel.getLibraryMaterials();
-    if (libraryMaterials != null) {
-      List<Material> materials = libraryMaterials.getMaterials();
-      for (Material material : materials) {
-        int index = getMaterialIndex(material.getId(), colladaModel);
-        boolean isUsed = false;
-        for (Mesh aliceMesh : aliceMeshes) {
-          if (aliceMesh.textureId.getValue() == index) {
-            isUsed = true;
-            break;
-          }
-        }
-        if (isUsed) {
-          Image image = getColladaImageForMaterial(material, colladaModel);
-          if (image == null) {
-            throw new ModelLoadingException("Error loading material " + material.getId() + ": No valid image found.");
-          }
-          BufferedImage bufferedImage;
-          try {
-            File imageFile = resolveTextureFileName(image.getInitFrom(), rootPath);
-            if (imageFile == null) {
-              throw new ModelLoadingException("Error loading texture: File '" + image.getInitFrom() + "' not found.");
-            }
-            bufferedImage = ImageUtilities.read(imageFile);
-            if (bufferedImage == null) {
-              throw new ModelLoadingException("Error loading texture: File '" + image.getInitFrom() + "' not readable.");
-            }
-            bufferedImage = ImageUtilities.stretchToPowersOfTwo(bufferedImage);
-          } catch (IOException e) {
-            throw new ModelLoadingException("Error loading texture: " + image.getInitFrom() + " not found.", e);
-          } catch (RuntimeException e) {
-            throw new ModelLoadingException("Error loading texture: " + image.getInitFrom() + ".\n" + e.getMessage(), e);
-          }
-          TexturedAppearance m_sgAppearance = new TexturedAppearance();
-          m_sgAppearance.diffuseColorTexture.setValue(getAliceTexture(bufferedImage));
-          m_sgAppearance.textureId.setValue(index);
-          textureAppearances.add(m_sgAppearance);
-        } else {
-          modelLoadingLogger.log(Level.WARNING, "Loading materials: Skipping unreferenced material " + material.getId());
-        }
-      }
-    }
-    if (textureAppearances.isEmpty()) {
-      throw new ModelLoadingException("No supported materials found. Alice models must have a texture.");
-    }
-    return textureAppearances;
-  }
-
-  private static File resolveTextureFileName(String textureFileName, File rootPath) {
-    //Strip URI info. This seems to be inconsistent with java.net.URI and therefore it's easier to discard it
-    if (textureFileName.startsWith("file://")) {
-      textureFileName = textureFileName.substring(7);
-    } else if (textureFileName.startsWith("file:///")) {
-      textureFileName = textureFileName.substring(8);
-    }
-
-    //Check for texture file
-    File textureFile = new File(textureFileName);
-    if (textureFile.exists()) {
-      return textureFile;
-    }
-    //If texture file isn't found, check in the root directory (the directory where the collada file is loaded from)
-    File localFile = new File(rootPath, textureFile.getName());
-    if (localFile.exists()) {
-      return localFile;
-    }
-    return null;
-  }
-
-  private Image getColladaImageForMaterial(Material material, Collada colladaModel) {
-    InstanceEffect ie = material.getInstanceEffect();
-    Effect effect = colladaModel.findEffect(ie.getUrl());
-    if (effect == null) {
-      modelLoadingLogger.warning("Error loading material '" + material.getId() + "': No effect found for url '" + ie.getUrl() + "'");
-      return null;
-    }
-    if (effect.getEffectMaterial() == null) {
-      modelLoadingLogger.warning("Error loading material '" + material.getId() + "': No effect material found for '" + effect.getId() + "'");
-      return null;
-    } else if (effect.getEffectMaterial().getDiffuse() == null) {
-      modelLoadingLogger.warning("Error loading material '" + material.getId() + "': No diffuse value found for effect '" + effect.getId() + "'");
-      return null;
-    } else if (effect.getEffectMaterial().getDiffuse().getTexture() == null) {
-      modelLoadingLogger.warning("Error loading material '" + material.getId() + "': No diffuse texture found for effect '" + effect.getId() + "'");
-      return null;
-    } else if (effect.getEffectMaterial().getDiffuse().getTexture().getTexture() == null) {
-      modelLoadingLogger.warning("Error loading material '" + material.getId() + "': No diffuse texture value found for effect '" + effect.getId() + "'");
-      return null;
-    }
-    String textureId = effect.getEffectMaterial().getDiffuse().getTexture().getTexture();
-
-    NewParam textureParam = effect.findNewParam(textureId);
-    while (textureParam != null) {
-      if (textureParam.getSurface() != null) {
-        textureId = textureParam.getSurface().getInitFrom();
-        textureParam = null;
-        break;
-      } else if (textureParam.getSampler2D() != null) {
-        textureParam = effect.findNewParam(textureParam.getSampler2D().getSource());
-      } else {
-        textureParam = null;
-      }
-    }
-
-    return colladaModel.findImage(textureId);
   }
 
   private Collada readColladaModel() throws ModelLoadingException {
@@ -559,7 +177,7 @@ public class JointedModelColladaImporter {
     if (scene == null) {
       throw new ModelLoadingException("Error processing model: No scene found.");
     }
-    //Find and build the skeleton first
+
     Node rootNode = findRootNode(scene.getNodes());
     Joint aliceSkeleton = null;
     if (rootNode != null) {
@@ -571,8 +189,8 @@ public class JointedModelColladaImporter {
       }
     }
 
-    //Find and build meshes (both static and weighted) from the collada model
-    List<Mesh> aliceMeshes = createAliceMeshesFromCollada(colladaModel);
+    ColladaImportMeshBuilder meshBuilder = new ColladaImportMeshBuilder(orientation, modelLoadingLogger);
+    List<Mesh> aliceMeshes = meshBuilder.createAliceMeshesFromCollada(colladaModel);
     if (aliceMeshes.isEmpty()) {
       throw new ModelLoadingException("Error processing model: No valid meshes found.");
     }
@@ -581,12 +199,10 @@ public class JointedModelColladaImporter {
     skeletonVisual.setName(baseFileName(colladaModelFile.getName()));
     skeletonVisual.frontFacingAppearance.setValue(new SimpleAppearance());
     skeletonVisual.skeleton.setValue(aliceSkeleton);
-    List<Mesh> aliceGeometry = new ArrayList<Mesh>();
-    List<WeightedMesh> aliceWeightedMeshes = new ArrayList<WeightedMesh>();
-    //Loop through the meshes and divide them into lists of regular meshes and weighted meshes
+    List<Mesh> aliceGeometry = new ArrayList<>();
+    List<WeightedMesh> aliceWeightedMeshes = new ArrayList<>();
     for (Mesh mesh : aliceMeshes) {
       if (mesh instanceof WeightedMesh weightedMesh) {
-        //Link the weighted mesh to the skeleton
         weightedMesh.skeleton.setValue(aliceSkeleton);
         aliceWeightedMeshes.add(weightedMesh);
       } else {
@@ -602,7 +218,8 @@ public class JointedModelColladaImporter {
       skeletonVisual.scale(extraScale);
     }
 
-    List<TexturedAppearance> sgTextureAppearances = createAliceMaterialsFromCollada(colladaModel, rootPath, aliceMeshes);
+    ColladaImportMaterialLoader materialLoader = new ColladaImportMaterialLoader(modelLoadingLogger);
+    List<TexturedAppearance> sgTextureAppearances = materialLoader.createAliceMaterialsFromCollada(colladaModel, rootPath, aliceMeshes);
     skeletonVisual.textures.setValue(sgTextureAppearances.toArray(new TexturedAppearance[sgTextureAppearances.size()]));
 
     UtilitySkeletonVisualAdapter skeletonVisualAdapter = new UtilitySkeletonVisualAdapter();
@@ -631,58 +248,5 @@ public class JointedModelColladaImporter {
       return  fileName;
     }
     return fileName.substring(0, pos);
-  }
-
-  /*
-   *  Convenience methods for debugging
-   */
-
-  private static void printGeometryInfo(Geometry geometry) {
-    float[] normalData = geometry.getMesh().getNormalData();
-    int normalCount = normalData.length / 3;
-    float[] vertexData = geometry.getMesh().getPositionData();
-    int vertexCount = vertexData.length / 3;
-    float[] uvData = geometry.getMesh().getTexCoordData();
-    int uvCount = uvData.length / 2;
-    Triangles tris = (Triangles) geometry.getMesh().getPrimitives().getFirst();
-    int triCount = tris.getCount();
-
-    System.out.println("Tris:  " + triCount + ", normals: " + normalCount + ", vertices: " + vertexCount + ", uvs: " + uvCount);
-  }
-
-  private static void printJoints(Joint j, String indent) {
-    System.out.println(indent + "Joint " + j.jointID.getValue());
-    PrintUtilities.print(indent + "    local transform: ", j.localTransformation.getValue().translation(), j.localTransformation.getValue().orientation());
-    System.out.println();
-    AffineMatrix4x4 absoluteTransform = j.getAbsoluteTransformation();
-    PrintUtilities.print(indent + " absolute transform: ", absoluteTransform.translation(), absoluteTransform.orientation());
-    System.out.println();
-    for (int i = 0; i < j.getComponentCount(); i++) {
-      Component comp = j.getComponentAt(i);
-      if (comp instanceof Joint joint) {
-        printJoints(joint, indent + "  ");
-      }
-    }
-  }
-
-  private static void printWeightInfo(WeightInfo wi) {
-    for (Entry<String, InverseAbsoluteTransformationWeightsPair> entry : wi.getMap().entrySet()) {
-      InverseAbsoluteTransformationWeightsPair iatwp = entry.getValue();
-      Point3 t = iatwp.getInverseAbsoluteTransformation().translation();
-      OrthogonalMatrix3x3 o = iatwp.getInverseAbsoluteTransformation().orientation();
-
-      System.out.println(entry.getKey() + ":");
-      System.out.println(" inverse transform = (" + t.x() + ", " + t.y() + ", " + t.z() + "), [[" + o.right().x() + ", " + o.right().y() + ", " + o.right().z() + "], [" + o.up().x() + ", " + o.up().y() + ", " + o.up().z() + "], [" + o.backward().x() + ", " + o.backward().y() + ", " + o.backward().z() + "]]");
-      List<Float> weights = new ArrayList<Float>();
-      List<Integer> indices = new ArrayList<Integer>();
-      InverseAbsoluteTransformationWeightsPair.WeightIterator weightIterator = iatwp.getIterator();
-      while (weightIterator.hasNext()) {
-        indices.add(weightIterator.getIndex());
-        weights.add(weightIterator.next());
-      }
-      System.out.println("  weight count = " + weights.size());
-      System.out.println("  weights: " + weights);
-      System.out.println(" indices: " + indices);
-    }
   }
 }
